pyproject.toml

[project]
name = "backend_PainelPesquisaEConsultoria"
version = "0.1.0"
description = ""
authors = [{ name = "dudubarbaro", email = "dudubarbaro" }]
dependencies = [
    "djangorestframework>=3.14.0",
    "django>=4.2.7",
    "setuptools>=68.2.2",
    "drf-spectacular>=0.26.5",
    "djangorestframework-simplejwt>=5.3.0",
    "pandas>=2.1.2",
    "django-cors-headers>=4.3.0",
    "python-dotenv>=1.0.0",
    "netifaces>=0.11.0",
    "django-filter>=23.3",
    "openpyxl>=3.1.2",
    "django-import-export>=3.3.3",
    "ta-lib>=0.4.28",
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.scripts]
pre_dev = "python ./scripts/set_my_ip.py"
# dev = "python manage.py runserver 0.0.0.0:19003"
dev = "python manage.py runserver"
runserver = "python manage.py runserver"
createsuperuser = "python manage.py createsuperuser"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"
shell = "python manage.py shell"
test = "python manage.py test"
startapp = "python manage.py startapp {args}"
loaddata = "python manage.py loaddata {args}"
dumpdata = "python manage.py dumpdata {args}"

[tool.pdm.dev-dependencies]
dev = ["isort>=5.12.0", "black>=23.10.1"]

[tool.black]
line-length = 120

[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"
